requirements.txt
-i https://pypi.org/simple/
certifi==2020.12.5
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
dnspython==1.16.0
Flask==1.1.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-dotenv==0.12.0
requests==2.27.1
six==1.15.0
urllib3==1.26.5
Werkzeug==1.0.1
